+++ src/Client/UI/Components/Toaster.ts
@@ -1,52 +1,7 @@
 import {Context} from "../Context.js";
 import {createDocumentFragment} from "../ElementExtensions.js";
 import {Position, toCss} from "../Position.js";
-
-/**
- * Represents a notification.
- */
-export interface IToast {
-
-	/**
-	 * Value indicating whether to apply a fade transition.
-	 */
-	animation?: boolean;
-
-	/**
-	 * Value indicating whether to automatically hide the toast.
-	 */
-	autoHide?: boolean;
-
-	/**
-	 * The title displayed in the header.
-	 */
-	caption: string;
-
-	/**
-	 * The child content displayed in the body.
-	 */
-	childContent: DocumentFragment|string;
-
-	/**
-	 * The default contextual modifier.
-	 */
-	context?: Context;
-
-	/**
-	 * The culture used to format the relative time.
-	 */
-	culture?: Intl.Locale;
-
-	/**
-	 * The delay, in milliseconds, to hide the toast.
-	 */
-	delay?: number;
-
-	/**
-	 * The icon displayed next to the caption.
-	 */
-	icon?: string|null;
-}
+import type {IToast} from "./Toast.js";
 
 /**
  * Manages the notifications.