+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version [0.17.0](https://github.com/cedx/base/compare/v0.16.0...v0.17.0)
+- Added the `UI.Components.IDialogButton` interface.
 - Added the `UI.Components.MessageBox` component.
 
 ## Version [0.16.0](https://github.com/cedx/base/compare/v0.15.3...v0.16.0)